Set up ZoteroConnector extension
This extension will allow Habele Institute to import references from Zotero,
along with the associated attachments. The references and attachments can be
imported manually by accounts with the `zotero-import` permission (by default,
granted to administrators) using the special page `Special:ZoteroImporter`, or
automatically via a maintenance script.

References are created in a new namespace, named "Zotero reference"; editing in
that namespace is restricted to those with the `edit-zotero-reference`
permission, which no one is assigned by default. It is expected that those
pages will only be created and updated by this extension and not need to be
edited manually. Reference pages are named based on their Zotero key.

Attachments are uploaded by URL to the file namespace, again named based on
their Zotero key. A new parser function, `#ZOTERO_FILE_VERSION`, is added to
allow storing and tracking the version of the attachment currently present on
the wiki; this parser function is only valid in the NS_FILE namespace, and
should only be used for Zotero attachment files. For the uploading to work, the
wiki must be configured to allow uploading files by URL, and to allow uploading
PDF files.

Test coverage is fairly minimal at the moment, but manual testing can be done
by running the import script and confirming the results.

HI1-2