Set up SemanticMediawiki integration
* On the reference pages, use SemanticMediaWiki's `#set` parser function to add
details of the authors, editors, publication, and publisher to the reference
page.
* When building up the citation templates, for the creators (authors and
editors) add a link parameter indicating that their names should be links to
`Special:Browse/{name}` so that the references that use the given author/editor
can be explored.
* When building up the citation templates, replace the display of the publisher
and publication with a similar link to `Special:Browse`.
* Add a PageExchange/PagePort package to easily create the various
SemanticMediaWiki properties that the extension uses.

HI1-3

---------

Co-authored-by: Daniel Scherzer <daniel@wikiteq.com>