Retry attachment imports if not too big
If any of the attachments fail to import for a reason other than being too big,
they will now be retried once in the hopes that the error was just a fluke.
Also include some tests for this behavior and the non-dry-mode importing of
attachments, uses a lot of mocking so not the nicest but now at least the code
is covered.

HI1-26